+++ src/azure/cosmos/request_response.rs
@@ -1,6 +1,7 @@
 use azure::cosmos::database::Database;
+use azure::cosmos::collection::Collection;
 
-#[derive(Deserialize)]
+#[derive(Deserialize, Debug)]
 #[allow(dead_code)]
 pub struct ListDatabasesResponse {
     _rid: String,
@@ -10,7 +11,17 @@ pub struct ListDatabasesResponse {
     pub count: u32,
 }
 
-#[derive(Serialize)]
+#[derive(Serialize, Debug)]
 pub struct CreateDatabaseRequest<'a> {
     pub id: &'a str,
+}
+
+#[derive(Deserialize, Debug)]
+#[allow(dead_code)]
+pub struct ListCollectionsResponse {
+    _rid: String,
+    #[serde(rename = "DocumentCollections")]
+    pub collections: Vec<Collection>,
+    #[serde(rename = "_count")]
+    pub count: u32,
 }