+++ src/azure/storage/blob/mod.rs
@@ -44,7 +44,9 @@ use azure::core::range::Range;
 use azure::core::ba512_range::BA512Range;
 use azure::core::incompletevector::IncompleteVector;
 
-use mime::Mime;
+//use mime::Mime;
+
+use hyper::mime::Mime;
 
 use hyper::status::StatusCode;
 use hyper::header::{Headers, ContentType, ContentLength, LastModified, ContentEncoding,
@@ -428,27 +430,27 @@ impl Blob {
                 if r.is_none() {
                     return Err(AzureError::InputParametersError("cannot use put_blob with \
                                                                  BlockBlob without a Read"
-                                                                        .to_owned()));
+                                                                    .to_owned()));
                 }
             }
             BlobType::PageBlob => {
                 if r.is_some() {
                     return Err(AzureError::InputParametersError("cannot use put_blob with \
                                                                  PageBlob with a Read"
-                                                                        .to_owned()));
+                                                                    .to_owned()));
                 }
 
                 if self.content_length % 512 != 0 {
                     return Err(AzureError::InputParametersError("PageBlob size must be aligned \
                                                                  to 512 bytes boundary"
-                                                                        .to_owned()));
+                                                                    .to_owned()));
                 }
             }
             BlobType::AppendBlob => {
                 if r.is_some() {
                     return Err(AzureError::InputParametersError("cannot use put_blob with \
                                                                  AppendBlob with a Read"
-                                                                        .to_owned()));
+                                                                    .to_owned()));
                 }
             }
         }